% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfast_interactive.R
\name{bfast_interactive}
\alias{bfast_interactive}
\title{Compute BFAST interactively}
\usage{
bfast_interactive(data_layers, obs_per_year, display = NULL,
  add_layers = NULL, cell_number = NULL, impute,
  nodata_threshold = c(0.05, 4), return_bfast = FALSE, ...)
}
\arguments{
\item{data_layers}{A RasterBrick or RasterStack object. A RasterBrick is
preferable to reduce processing time.}

\item{obs_per_year}{A number indicating the number of observations in the
time series per year. For example, this argument should be set to 12 for
monthly data. For MODIS 16-day aggregates, this should be set to 23
(365/16, rounded to nearest integer.)}

\item{display}{A number or a RasterLayer object. If \code{display} is a
number, the corresponding layer from \code{data_layers} will be plotted for
the cell selection. If \code{display} is a RasterLayer object, it will be
plotted for cell selection.}

\item{add_layers}{An optional list of Spatial objects to be plotted on top of
\code{display}.}

\item{cell_number}{A number indicating the cell number through which
\code{bfast()} should be run from \code{data_layers}. Ignored if
\code{display} is supplied.}

\item{impute}{A logical indicating whether the function should linearly
interpolate through time series with missing data if the time series does
not exceed the thresholds defined by \code{nodata_threshold}.}

\item{nodata_threshold}{A numeric vector of length 2. The first entry is the
proportion of missing data points in each time series above which the
function will not execute and return \code{NA} for all fields. The second
entry is the number of consecutive missing data points above which the
function will not execute and return \code{NA} for all fields.}

\item{return_bfast}{A logical indicating if the BFAST results should be
returned.}

\item{...}{Arguments passed on to \code{bfast}.}
}
\value{
A bfast object (if \code{return_bfast=TRUE}) or nothing otherwise.
}
\description{
This function allows the selection of a cell by clicking on a raster plot,
and then executes \code{bfast::bfast()} for the cell's time series. The
function then runs and plots \code{bfast()} on the time series.
}
\examples{
\dontrun{
bfast_interactve(mod.brick, monthly=TRUE, display=raster("C:/Desktop/Mojave.tif", add_layers=list(boundary), impute=TRUE, nodata_threshold=c(0.05, 3))
}
}
