% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_points.R
\name{create_points}
\alias{create_points}
\title{Create an ESRI shapefile containing point features indexed to cell number}
\usage{
create_points(template, boundary = NULL, output_file = NULL)
}
\arguments{
\item{template}{A Raster object to be convert into point features.
Alternatively, an XML file generated by \code{create_raster_metadata()}.}

\item{boundary}{An optional SpatialPolygons object to subset the point
features after they have been generated according to the the grid from
\code{template}. It should be in the same projection as \code{template}.}

\item{output_file}{An optional path to a .shp file to which the shapefile
will be written. If this argument is not supplied, the features will be
returned as a SpatialPointsDataFrame object.}
}
\value{
A SpatialPointsDataFrame object or nothing (with a side-effect of
  writing to a file).
}
\description{
This function converts a raster into a shapefile with point features
corresponding to each cell. Each point will have an attribute for its cell
number.
}
\details{
If an output file is specified, this function will write an ESRI shapefile to
that file. The cell number will appear as the "layer" field in the attribute
table.
}
\examples{
\dontrun{
create_points(raster("C:/Desktop/extent.tif"), "C:/Desktop/shapefiles/all_points.shp")
}
}
