% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_raster_metadata.R
\name{create_raster_metadata}
\alias{create_raster_metadata}
\title{Create raster XML metadata file}
\usage{
create_raster_metadata(data_layers, output_file)
}
\arguments{
\item{data_layers}{A Raster object for which metadata will be created.}

\item{output_file}{A character file path to the XML file to which the
metadata will be written.}
}
\value{
Nothing. This function has a side-effect of writing to a file.
}
\description{
This function writes an XML file containing the information on the extent,
projection, and dimensions of the input dataset. This is designed to be an
alternative for the \code{template} argument for other functions.
}
\details{
The information in the XML file does not create an exactly identical raster
in terms of extent and resolution compared to the original input raster, most
likely because of infinite precision issues. In other words, testing the
extent and resolution of a raster created from the XML file and the original
raster with \code{identical()} yields \code{FALSE}. However, testing them
with \code{all.equal()}, which allows for small numerical differences, yields
\code{TRUE}. It is likely that this is not a major problem because the
\code{raster} package still performs map algebra without issue. However, this
has not yet been tested in ArcGIS (and may apprently be remedied with the
Snap Raster environment setting if it is a problem). For all display
purposes, the rasters may be considered as exactly identical.
}
\examples{
\dontrun{
create_raster_metadata(mod.brick, "C:/Desktop/metadata.xml")
}
}
