% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_MODIS_date.R
\name{generate_MODIS_date}
\alias{generate_MODIS_date}
\title{Generate a date vector for MODIS aggregate products}
\usage{
generate_MODIS_date(start_date, end_date, interval, subset = NULL,
  inverse = FALSE)
}
\arguments{
\item{start_date}{A Date or character of format YYYY-MM-DD for the starting
date of the vector.}

\item{end_date}{A Date or character of format YYYY-MM-DD for the ending date
of the vector.}

\item{interval}{The time in days between each entry in the vector. For
example, \code{interval} should be 16 for 16-day data.}

\item{subset}{A numeric vector of length 2 or 4. If the length is 2,
\code{subset} defines a yearly subset of the date vector by Julian date,
inclusive. For example, \code{subset=c(50, 150)} would define a subset to
keep only the dates every year whose Julian dates fall within that range.
If the length is 4, \code{subset} is defined in terms of a starting month
and date and an ending month and date (e.g. \code{c(2, 4, 6, 27)} refers to
a subset starting February 4 and ending June 27 for each year).}

\item{inverse}{A logical indicating if the subset defined by \code{subset}
should be inverted to find only dates \strong{outside} the subset.}
}
\value{
A vector of class Date containing dates corresponding to MODIS data
  dates.
}
\description{
This function generates a date vector that correponds to the observation
dates for MODIS aggregated products (e.g. 8-day and 16-day aggregagtes).
}
\details{
This function assumes that the dates for a given year begin on January 1 and
increment by \code{interval}. This pattern matches that of 8-day and 16-day
aggregate products, but tests against other products have not been performed.
}
\examples{
\dontrun{
generate_MODIS_date("2000-02-28", "2009-07-02", 8)
}
}
