% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_date.R
\name{monthly_date}
\alias{monthly_date}
\title{Create a monthly date vector}
\usage{
monthly_date(st_month, st_year, ed_month, ed_year, date_number = 1)
}
\arguments{
\item{st_month}{A number indicating the month of the starting date.}

\item{st_year}{A number indicating the year of the starting date.}

\item{ed_month}{A number indicating the month of the ending date.}

\item{ed_year}{A number indicating the year of the ending date.}

\item{day_numer}{A number indicating the dummy date number that should be set
for each date.}
}
\value{
A vector of class Date spanning the start and end dates by month.
}
\description{
This function produces a vector of dates that increment by month.
}
\details{
The output of this function is most useful for passing as an argument to
\code{bfastSpatial()}, which requires a date vector.
}
\examples{
\dontrun{
monthly_date(2, 2000, 5, 2018)
}
}
