% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcf_calc.R
\name{pcf_calc}
\alias{pcf_calc}
\title{Compute pair correlation function for breakpoint distribution}
\usage{
pcf_calc(bfast_in, level = 0.05, boundary, template, resample_dist = NULL,
  max.dist = NULL, output_directory = NULL, mc.cores = 1)
}
\arguments{
\item{bfast_in}{A data frame generated by \code{bfastSpatial()}.}

\item{level}{A number indicating the significance level for the pointwise
Monte Carlo test. See the documentation for \code{spatstat::envelope()} for
more details. If \code{level=NA}, no envelope will be simulated.}

\item{boundary}{A SpatialPolygons object defining the boundary of the PCF
analysis. If this argument is not supplied, all records in \code{bfast_in}
will be used.}

\item{template}{A Raster object with the same resolution, extent, and
projection as the raster dataset used to generate \code{bfast_in}.
Alternatively, an XML file generated by \code{create_raster_metadata()}.}

\item{resample_dist}{A number indicating the square dimension to which the
breakpoint pattern raster should be resampled. This argument is in the same
units as the projection of \code{template}.}

\item{max.dist}{The maximum radius distance to which the PCF calculation
should be performed. If not supplied, the default is the diagonal distance
across \code{boundary}.}

\item{output_directory}{An optional character file path to a directory to
which the PCF calculations for each year and month should be written in
\code{.rds} format. The radius distances for which the PCF was computed
will also be written here.}

\item{mc.cores}{A numeric indicating the number of cores to be used in
parallel computing.}
}
\value{
A list of length 2. The first item is a data frame summarizing the
  PCF results. The second item is the vector of radius distances.
}
\description{
This function computes the pair correlation function (PCF) for the point
pattern of breakpoints. The PCF estimates \emph{g} for a given radius
distance \emph{r}, where \emph{g} represents the degree of clustering.
}
\details{
This function is a convenient wrapper for \code{spatstat::pcf()} and
\code{spatstat::envelope()}. Most default values in \code{pcf} and
\code{envelope} are accepted. However, the translation correction is used
here.

If \code{level} is not set to \code{NA}, a simulation envelope will be
computed according to the number of simulations needed to achieve
\code{level}. However, the simulation process is very time-consuming so
setting \code{level=NA} may be best choice for reasonable processing times.

The returned data frame has the following fields: \enumerate{ \item
\code{year}: The year of the observation. \item \code{month}: The month of
the observation. \item \code{type}: The type of statistic for \emph{g}. \item
\code{g}: The value of \emph{g} for the given \code{type}. \item \code{dist}:
The radius distance. \item \code{upper.bound, lower.bound}: The upper/lower
bound of the simulation envelope at distance \code{dist}. \item \code{nsim}:
The number of simulations to produce the envelope. \item \code{sig.level}:
The significance level of the pointwise Monte Carlo test. }
}
\examples{
\dontrun{
pcf_calc(bf_df, NA, rgdal::readOGR(dsn="C:/Desktop/shapefiles", layer="Mojave"), template, 2000, "C:/Desktop/PCF_files", 6)
}
}
