% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_PR.R
\name{process_PR}
\alias{process_PR}
\title{Process MODIS raster layers for pixel reliability}
\usage{
process_PR(data_layers, pixel_reliability, selected_PR, basename,
  output_directory, bylayer = FALSE, clean = FALSE, mc.cores = 1)
}
\arguments{
\item{data_layers}{A RasterStack or RasterBrick object containing the MODIS
data. A RasterBrick object is preferable to reduce processing time.}

\item{pixel_reliability}{A RasterStack or RasterBrick object containing the
MODIS pixel reliability layers. A RasterBrick object is preferable to
reduce processing time.}

\item{selected_PR}{A list of length 1 containing a named vector. This name
must be either \code{"include"} or \code{"exclude"}. See Details for more
information.}

\item{basename}{A character that will be prepended to all file names to be
written.}

\item{output_directory}{A character file path to a directory where all files
will be written.}

\item{bylayer}{A logical indicating if the individual raster outputs should
remain separate (\code{TRUE}) or combined into a single multilayered raster
file (\code{FALSE}).}

\item{clean}{A logical indicating if the individual raster outputs should be
deleted once the multilayered raster has been created. Ignored if
\code{bylayer=TRUE}.}

\item{mc.cores}{A numeric indicating the number of cores to be used in
parallel computing.}
}
\value{
Nothing. This function has the side-effect of writing to files.
}
\description{
This function sets values in a MODIS raster dataset to \code{NA} according to
pixel reliability values. All raster outputs will be written to GeoTIFF
files.
}
\details{
NB: If you are using \code{MODIStsp} to derive pixel reliability, please note
that it seems that pixel reliability 0 corresponds to \code{NA}.

Since this function matches the layers in \code{data_layers} and
\code{pixel_reliability} by position, please ensure that all raster layers in
both datasets are sorted in the corrected order. For example, the first layer
in both datasets should correspond to the same time point.

Note that writing to a single multilayered raster means that layer names will
not be preserved. That is to say, all layers will simply be named according
to the file name with a numerical index appended when the multilayered
GeoTIFF is read back into R. The drawback is that any date information
contained in the layer name will be lost. Thus, it is advisable to keep the
individual raster files so that one can reference back to the date
information.

The \code{selected_PR} argument requires list containing a single named
numerical vector. The vector must be named either \code{"include"} or
\code{"exclude"} to determine the pixel reliability selection. If the vector
is named \code{"include"}, the numbers in the vector will be
\strong{included} in the pixel reliability values whose corresponding cells
will be filtered out (e.g. set to \code{NA}). If the vector is named
\code{"exclude"}, the numbers in the vector will be \strong{excluded} from
the pixel reliability values that will be used to filter out cells. Any pixel
reliability value that is not in the exclude vector will be used to filter
out cells.
}
\examples{
\dontrun{
process_PR(mod.brick, pr, list(exclude=c(NA, 1)), "MODIS_EVI", "C:/Desktop/PR_processed", mc.cores=6)
}
}
