% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reaggregate_by_function.R
\name{reaggregate_by_function}
\alias{reaggregate_by_function}
\title{Re-aggregate raster layers according to function}
\usage{
reaggregate_by_function(data_layers, fun, dates, basename, output_directory,
  bylayer = FALSE, clean = FALSE, mc.cores = 1)
}
\arguments{
\item{data_layers}{A RasterStack or RasterBrick object containing the MODIS
data. A RasterBrick object is preferable to reduce processing time.}

\item{fun}{A function that returns a single value and has an \code{na.rm}
argument, which will be set to \code{TRUE}. Examples are \code{mean, sum,
max, median}, and \code{min}.}

\item{dates}{A vector of class Date or a number. If \code{dates} is a date
vector, the data will be re-aggregated to monthly data by \code{fun}. If
\code{dates} is a number, the data will be re-aggregated in groups
specified by the number. For example, if \code{dates=4} the data will be
reaggregated by grouping layers in groups of 4 and executing \code{fun}.}

\item{basename}{A character that will be prepended to all file names to be
written.}

\item{output_directory}{A character file path to a directory where all files
will be written.}

\item{bylayer}{A logical indicating if the individual raster outputs should
remain separate (\code{TRUE}) or combined into a single multilayered raster
file (\code{FALSE}).}

\item{clean}{A logical indicating if the individual raster outputs should be
deleted once the multilayered raster has been created. Ignored if
\code{bylayer=TRUE}.}

\item{mc.cores}{A numeric indicating the number of cores to be used in
parallel computing.}
}
\value{
Nothing. This function has a side-effect of writing files to a
  directory.
}
\description{
This function re-aggregates a multilayered raster dataset according to a
function (e.g. maximum, minimum, mean) and a time interval. All raster
outputs will be written to GeoTIFF files.
}
\details{
Note that writing to a single multilayered raster means that layer names will
not be preserved. That is to say, all layers will simply be named according
to the file name with a numerical index appended when the multilayered
GeoTIFF is read back into R. The drawback is that any date information
contained in the layer name will be lost. Thus, it is advisable to keep the
individual raster files so that one can reference back to the date
information.
}
\examples{
\dontrun{
reaggregate_by_function(data_layers, max, "MODIS_max", "C:/Desktop", mc.cores=6)
}
}
