% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_MODIS_by_date.R
\name{select_MODIS_by_date}
\alias{select_MODIS_by_date}
\title{Select MODIS files by date}
\usage{
select_MODIS_by_date(files, output_directory, output_filename = NULL,
  selection, inverse = FALSE, convention = "standard", bylayer = FALSE)
}
\arguments{
\item{files}{A character vector of files from which to select.}

\item{output_directory}{A character file path to a directory where files will
be written.}

\item{output_filename}{A character for the name of the output multilayered
raster. Ignored if \code{bylayer=TRUE}.}

\item{selection}{A numeric vector of length 2 for the selection range
(inclusive). The first value should be the date or Julian date of the start
of the selection range. The second value should be the date or Julian date
of the end of the selection range.}

\item{inverse}{A logical indicating if the selection range should be inverted
(e.g. select files that are outside the selection range.) If \code{TRUE},
the inverted selection range is exclusive.}

\item{convention}{A character. Set this argument to "standard" if the file
names are in the standard MODIS naming convention. Set this argument to
"MODIStsp" if the files have been processed using \code{MODIStsp} (and thus
follow its naming convention).}

\item{bylayer}{A logical indicating if the individual raster outputs should
remain separate (\code{TRUE}) or combined into a single multilayered raster
file (\code{FALSE}).}
}
\description{
This function copies or combines to a multilayer raster MODIS files that
match the specified date criteria. Dates may be specified using Julian date
or traditional date.
}
