% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_bfast.R
\name{subset_bfast}
\alias{subset_bfast}
\title{Subset a \code{bfastSpatial} data frame by location}
\usage{
subset_bfast(bfast_in, template, boundary, inverse = FALSE)
}
\arguments{
\item{bfast_in}{A data frame generated by \code{bfastSpatial()}.}

\item{template}{A Raster object with the same resolution, extent, and
projection as the raster dataset used to generate \code{bfast_in}.
Alternatively, an XML file generated by \code{create_raster_metadata()}.}

\item{boundary}{A SpatialPolygons object used to define the subset.}

\item{inverse}{A logical indicating if the spatial subset defined by
\code{boundary} should be inverted to keep records for cells
\strong{outside} of it.}
}
\value{
A data frame subsetted according to \code{boundary}.
}
\description{
This function subsets the data frame generated by \code{bfastSpatial} by
location.
}
\details{
This function is useful because the full data frame may be too large for easy
data manipulation. This function may also be used in conjunction with
\code{summary_stat} for zonal statistics.
}
\examples{
\dontrun{
subset_bfast(bf_df, raster(C:/Desktop/MODIS_NDVI.tif), bnd)
}
}
