% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_plot.R
\name{summary_plot}
\alias{summary_plot}
\title{Create summary statistic plots from \code{bfastSpatial} output by site}
\usage{
summary_plot(bfast_in, site, boundary = NULL, template = NULL)
}
\arguments{
\item{bfast_in}{A data frame generated by \code{bfastSpatial()}.}

\item{site}{A character giving the name of the site for which plots are
created. The only purpose of this argument is to add to the title of each
plot.}

\item{boundary}{An optional SpatialPolygons object used to define the subset.
If not supplied, the plot will be made based on all records in
\code{bfast_in}.}

\item{template}{An optional Raster object with the same resolution, extent,
and projection as the raster dataset used to generate \code{bfast_in}.
Alternatively, an XML file generated by \code{create_raster_metadata()}.
This argument must be supplied if \code{boundary} is supplied; otherwise,
it is ignored.}
}
\value{
A list of 7 ggplot2 plots.
}
\description{
This function returns a list of plots that summarizes the information from
\code{bfastSpatial} in plots for a given region.
}
\details{
The plots are: \enumerate{ \item Bar chart of frequency of number of
breakpoints \item Histogram of shifts \item Histogram of slopes \item
Histogram of standard deviations of residuals \item Proportions of all
positive or all negative slopes \item Histogram of breakpoints by date \item
Histogram of slope for cells without significant structural change \item
Histogram of standard deviations of residuals for cells without significant
structural change }

The \code{plot_grid()} function in the \code{cowplot} package arranges plots
generated from \code{ggplot2} in a grid if a single window view is desired.
}
\examples{
\dontrun{
summary_plot(bf_df, "Mojave", raster("C:/Desktop/Mojave.tif"), 1)
}
}
