% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vg_calc.R
\name{vg_calc}
\alias{vg_calc}
\title{Perform variogram analysis on \code{bfastSpatial()} shift and slope
statistics}
\usage{
vg_calc(bfast_in, boundary = NULL, template, cutoff_dist = NULL,
  mc.cores = 6)
}
\arguments{
\item{bfast_in}{A data frame generated by \code{bfastSpatial()}.}

\item{boundary}{A SpatialPolygons object defining the boundary of the
variogram analysis. If this argument is not supplied, all records in
\code{bfast_in} will be used.}

\item{template}{A Raster object with the same resolution, extent, and
projection as the raster dataset used to generate \code{bfast_in}.
Alternatively, an XML file generated by \code{create_raster_metadata()}.}

\item{cutoff_dist}{An optional data frame containing the output of
\code{pcf_calc()} for the same study area.}

\item{mc.cores}{A numeric indicating the number of cores to be used in
parallel computing.}
}
\value{
A data frame with variogram results.
}
\description{
This function uses the \code{gstat} package to compute variogram statistic
for the shift and slope output from \code{bfastSpatial} output. This function
uses the exponential variogram model.
}
\details{
This function returns a data frame with the following fields: \enumerate{
\item \code{year}: The year of the observation. \item \code{month}: The month
of the observations. \item \code{stat}: The BFAST statistics (\code{"slope"}
or \code{"shift"}) \item \code{model}: The variogram model (\code{"Nug"} or
\code{"Exp"}) \item \code{psill}: The partial sill. \item \code{range}: The
range parameter of the model. Only valid for Exp model. \item
\code{effective_range}: The effective range of the model (within 5% of the
sill). Only valid for Exp model. \item \code{SSerr}: The sum of squared
errors of the model from the sample variogram. \item \code{n}: The number of
points in the sample variogram. }

If the variogram model is singular or does not converge after 200 iterations,
no model is fit and \code{NA}s are returned for the entire row. Since this
function tries to fit a large number of models, it may fail to fit a
substantial proportion of models if the sample variograms do not match the
model well (or require transformation).
}
\examples{
\dontrun{
vg_calc(bf_df, boundary=rgdal::readOGR(dsn="C:/Desktop/shapefiles", layer="Mojave"), template="C:/Desktop/raster_metadata.xml", cutoff_dist=pcf_results, mc.cores=6)
}
}
