% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MODIStsp_date.R
\name{MODIStsp_date}
\alias{MODIStsp_date}
\title{Create a date vector from \code{MODIStsp}-generated RasterStack}
\usage{
MODIStsp_date(stack)
}
\arguments{
\item{stack}{A RasterStack object generated by \code{MODIStsp()}.
Alternatively, a character vector of file names (\strong{not} file paths)
generated by \code{MODIStsp} from which dates should be extracted.}
}
\value{
A vector of class Date containing the dates for each raster layer.
}
\description{
This function creates a vector of class Date that corresponds to the
observations in MODIS data processed by \code{MODIStsp}.
}
\details{
The output of this function is most useful for passing as an argument to
\code{bfastSpatial()}, which requires a date vector.
}
\examples{
\dontrun{
MODIStsp_date(raster_ts)
}
}
